test(config): move and extend chromium user data tests
Move Settings tests into the unit test module and extend validation
coverage for chromium_user_data_dir. Add imports and os usage to the
unit test module, and restore pydantic_settings reload handling.

- relocate tests from tests/core/test_config.py into
  tests/unit/core/test_config.py (merged into existing module)
- import os and Settings where needed; reorder imports to avoid
  circular reload issues
- add paramerized blank-value test to coerce whitespace to None
- add test to normalize relative paths to absolute paths
- add test to ensure fallback to original value if abspath raises
  OSError

These changes centralize configuration tests with unit tests and
increase robustness of Settings validators by verifying edge cases.
