+++ tests/unit/services/browser/actions/test_scroll.py
