+++ tests/unit/services/common/browser/test_cookie_repository.py
