+++ tests/unit/services/common/browser/test_cookies_manager.py
