+++ tests/unit/services/tiktok/search/actions/test_ui_controls.py
